+++ Automate/Framework/Machines/Buildings/ShippingBinMachine.cs
@@ -68,12 +68,12 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
         public override bool SetInput(IStorage input)
         {
             // get next item
-            ITrackedStack? tracker = input.GetItems().FirstOrDefault(p => p.Sample is SObject obj && obj.canBeShipped());
+            ITrackedStack? tracker = input.GetItems().FirstOrDefault(p => p.Sample.canBeShipped());
             if (tracker == null)
                 return false;
 
             // ship item
-            SObject item = (SObject)tracker.Take(tracker.Count)!;
+            Item item = tracker.Take(tracker.Count)!;
             var binList = (this.Location as Farm ?? Game1.getFarm()).getShippingBin(Game1.MasterPlayer);
             Utility.addItemToThisInventoryList(item, binList, listMaxSpace: int.MaxValue);