+++ LookupAnything/Framework/Lookups/Characters/FarmerSubject.cs
@@ -72,6 +72,14 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
             yield return new SkillBarField(I18n.Player_FishingSkill(), target.experiencePoints[SFarmer.fishingSkill], maxSkillPoints, skillPointsPerLevel);
             yield return new SkillBarField(I18n.Player_CombatSkill(), target.experiencePoints[SFarmer.combatSkill], maxSkillPoints, skillPointsPerLevel);
 
+            // custom skills
+            var spaceCore = this.GameHelper.SpaceCore;
+            if (spaceCore.IsLoaded)
+            {
+                foreach (string skill in spaceCore.ModApi.GetCustomSkills())
+                    yield return new SkillBarField(spaceCore.ModApi.GetDisplayNameOfCustomSkill(skill), spaceCore.ModApi.GetExperienceForCustomSkill(target, skill), maxSkillPoints, skillPointsPerLevel);
+            }
+
             // luck
             string luckSummary = I18n.Player_Luck_Summary(percent: (Game1.player.DailyLuck >= 0 ? "+" : "") + Math.Round(Game1.player.DailyLuck * 100, 2));
             yield return new GenericField(I18n.Player_Luck(), $"{this.GetSpiritLuckMessage()}{Environment.NewLine}({luckSummary})");