+++ LookupAnything/GameHelper.cs
@@ -11,6 +11,7 @@ using Pathoschild.Stardew.Common.Integrations.CustomFarmingRedux;
 using Pathoschild.Stardew.Common.Integrations.ExtraMachineConfig;
 using Pathoschild.Stardew.Common.Integrations.MultiFertilizer;
 using Pathoschild.Stardew.Common.Integrations.ProducerFrameworkMod;
+using Pathoschild.Stardew.Common.Integrations.SpaceCore;
 using Pathoschild.Stardew.Common.Items;
 using Pathoschild.Stardew.LookupAnything.Framework;
 using Pathoschild.Stardew.LookupAnything.Framework.Constants;
@@ -25,6 +26,7 @@ using StardewValley.Buildings;
 using StardewValley.Characters;
 using StardewValley.Extensions;
 using StardewValley.GameData.Crafting;
+using StardewValley.GameData.Crops;
 using StardewValley.GameData.FishPonds;
 using StardewValley.Locations;
 using StardewValley.Menus;
@@ -79,11 +81,14 @@ namespace Pathoschild.Stardew.LookupAnything
         /// <summary>The Custom Bush integration.</summary>
         public CustomBushIntegration CustomBush { get; }
 
+        /// <summary>The Extra Machine Config integration.</summary>
+        public ExtraMachineConfigIntegration ExtraMachineConfig { get; }
+
         /// <summary>The MultiFertilizer integration.</summary>
         public MultiFertilizerIntegration MultiFertilizer { get; }
 
-        /// <summary>The Extra Machine Config integration.</summary>
-        public ExtraMachineConfigIntegration ExtraMachineConfig { get; }
+        /// <summary>The SpaceCore integration to get info about its custom skills.</summary>
+        public SpaceCoreIntegration SpaceCore { get; }
 
 
         /*********
@@ -104,9 +109,10 @@ namespace Pathoschild.Stardew.LookupAnything
             this.BushBloomMod = new BushBloomModIntegration(modRegistry, monitor);
             this.CustomBush = new CustomBushIntegration(modRegistry, monitor);
             this.CustomFarmingRedux = new CustomFarmingReduxIntegration(modRegistry, monitor);
+            this.ExtraMachineConfig = new ExtraMachineConfigIntegration(modRegistry, monitor);
             this.MultiFertilizer = new MultiFertilizerIntegration(modRegistry, monitor);
             this.ProducerFrameworkMod = new ProducerFrameworkModIntegration(modRegistry, monitor);
-            this.ExtraMachineConfig = new ExtraMachineConfigIntegration(modRegistry, monitor);
+            this.SpaceCore = new SpaceCoreIntegration(modRegistry, monitor);
 
             this.ResetCache(monitor);
         }
@@ -199,6 +205,19 @@ namespace Pathoschild.Stardew.LookupAnything
 
                 select new KeyValuePair<string, bool>(obj.QualifiedItemId, Game1.player.basicShipped.ContainsKey(obj.ItemId))
             );
+        }
+
+        /// <summary>Get the first crop which returns this item has a harvest, if any.</summary>
+        /// <param name="itemId">The unqualified item ID to check.</param>
+        public static CropData? GetCropDataByHarvestItem(string itemId)
+        {
+            foreach (CropData crop in Game1.cropData.Values)
+            {
+                if (crop.HarvestItemId == itemId)
+                    return crop;
+            }
+
+            return null;
         }
 
         /// <summary>Get all items owned by the player.</summary>