+++ ChestsAnywhere/docs/release-notes.md
@@ -2,7 +2,9 @@
 
 # Release notes
 ## Upcoming release
+* Updated for Stardew Valley 1.6.9.
 * Improved dropdown positioning to reduce overlap with big chest UI.
+* Improved translations. Thanks to ForgottenZ (updated Chinese) and MakinDay (updated Italian)!
 
 ## 1.24.1
 Released 01 July 2024 for SMAPI 4.0.7 or later.