+++ DataLayers/docs/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* Updated for Stardew Valley 1.6.9.
 * Added support for Ultimate Fertilizer's multi-fertilizer feature.
 
 ## 1.17.2