+++ FastAnimations/docs/release-notes.md
@@ -1,6 +1,10 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Updated for Stardew Valley 1.6.9.
+* Improved translations. Thanks to MakinDay (updated Italian)!
+
 ## 1.13.3
 Released 07 July 2024 for SMAPI 4.0.7 or later.